requirements.txt
streamlit==1.41.1
pandas==2.2.3
plotly==6.0.0
numpy==2.2.2
wordcloud==1.9.4
matplotlib==3.10.0
seaborn==0.13.2
scikit-learn==1.6.1
python-dateutil==2.9.0.post0